+++ src/Modifiers.cs
@@ -1,0 +1,13 @@
+using System;
+
+namespace fierrof.ActionBar
+{
+    [Flags]
+    public enum Modifiers
+    {
+        None  = 0,
+        Ctrl  = 1,
+        Shift = 2,
+        Alt   = 4,
+    }
+}